requirements.txt
Babel==2.5.3
certifi==2018.1.18
chardet==3.0.4
Django==2.0.3
django-bootstrap3==9.1.0
django-localflavor==2.0
django-phonenumber-field==2.0.0
docopt==0.6.2
idna==2.6
mysqlclient==1.3.12
phonenumberslite==8.9.3
Pillow==5.0.0
pipreqs==0.4.9
pkg-resources==0.0.0
pytz==2018.3
requests==2.18.4
urllib3==1.22
yarg==0.1.9
